# Fix MaxUnavailableZones computed by PartitionInstanceRing.GetReplicationSetsForOperation()