@dimitarvdimitrov 2024-02-19
> we have some (or all) instances in the same zone

With the implementation in this PR would we end up querying both instances in the same zone or only one of them?