@pracucci 2024-02-19
> TestDefaultResultTracker_StartMinimumRequests_NoFailingRequests

I don't think it's related because the modified code doesn't touch the code run by that test at all, but I'm not in a hurry on this PR and tomorrow I will look at why `TestDefaultResultTracker_StartMinimumRequests_NoFailingRequests` is flaky.